random improvements idk
